tests: move some helper functions to util
